+++ types/core/api/SimBriefAeroflyApi.d.ts
@@ -1,6 +1,9 @@
 import { AeroflyFlight, AeroflyNavRouteWaypoint } from "@fboes/aerofly-custom-missions";
 import { SimBriefApi, SimBriefApiPayload, SimBriefApiPayloadAirport } from "./SimBriefApi.js";
+import { AeroflyAircraftService } from "../services/AeroflyAircraftService.js";
 export declare class SimBriefAeroflyApi extends SimBriefApi {
+    private aircraftService;
+    constructor(aircraftService: AeroflyAircraftService);
     fetchMission(username: string, flight: AeroflyFlight, useDestinationWeather?: boolean): Promise<void>;
     convertMission(simbriefPayload: SimBriefApiPayload, flight: AeroflyFlight, useDestinationWeather?: boolean): void;
     protected getWaypointsFromNavlog(simbriefPayload: SimBriefApiPayload): AeroflyNavRouteWaypoint[];