+++ types/core/services/AeroflyAirportService.d.ts
@@ -1,15 +1,10 @@
-export type AeroflyAirportSet = {
-    code: string;
-    name: string;
-    lat: number;
-    lon: number;
-};
+import type { AeroflyAirportCoordinatesObject } from "@fboes/aerofly-data/data/airport-coordinates-object.json";
 /**
  * Find detail data for Aerofly FS airports.
  * Interface to `@fboes/aerofly-data` JSON data.
  */
 export declare class AeroflyAirportService {
-    static getAllAirports(): AeroflyAirportSet[];
-    static getAirportByIcaoCode(icaoCodeAirport: string): AeroflyAirportSet | undefined;
+    getAllAirports(): AeroflyAirportCoordinatesObject[];
+    getAirportByIcaoCode(icaoCodeAirport: string): AeroflyAirportCoordinatesObject | undefined;
 }
 //# sourceMappingURL=AeroflyAirportService.d.ts.map